This commit was manufactured by cvs2svn to create tag 'VRJ_RELENG_2_0_0_RELEASE'.
git-svn-id: https://vrjuggler.googlecode.com/svn/tags/VRJ_RELENG_2_0_0_RELEASE@17859 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vrjuggler/test/Draw/OGL/cubes/cubesApp.cpp
@@ -191,7 +191,7 @@ void cubesApp::init()
    vprDEBUG(vprDBG_ALL, vprDBG_VERB_LVL) << "---------- cubes:App:init() ---------------"
                         << std::endl << vprDEBUG_FLUSH;
    std::vector<vrj::User*> users = mKernel->getUsers(); // Request user list
-   int num_users = users.size();
+   size_t num_users = users.size();
    vprASSERT(num_users > 0);      // Make sure that we actually have users defined
 
    UserData* new_user=NULL;
